requirements.txt
timefold-solver==1.10.0a0

